modified modals structure folder and composed new components needed
