+++ src/pages/blog_client/Home/constants/texts.constants.ts
