+++ src/pages/blog_client/Resources.tsx
