+++ src/pages/diagnoses/components/modals/modal_create/subsections/Biopsy.tsx
@@ -5,12 +5,12 @@ interface IProps {
   onClose: () => void;
   refModal: React.RefObject<HTMLDivElement>;
 }
-function ModalCytology({ onClose, refModal }: IProps): ReactElement {
+function Biopsy({ onClose, refModal }: IProps): ReactElement {
   return (
     <div className="fixed inset-0 z-20 flex w-full items-center justify-center bg-gray-400 bg-opacity-50 p-10 backdrop-blur-sm">
       <GeneralModal onClose={onClose} refModal={refModal}>
         <div className="flex h-full w-full flex-col items-start">
-          <h1 className="mx-4 pb-5 pt-3 text-3xl font-bold">Create Cytopathology</h1>
+          <h1 className="mx-4 pb-5 pt-3 text-3xl font-bold">Create Biopsy</h1>
           <div className="w-full p-2"></div>
         </div>
       </GeneralModal>
@@ -18,4 +18,4 @@ function ModalCytology({ onClose, refModal }: IProps): ReactElement {
   );
 }
 
-export default ModalCytology;
+export default Biopsy;
