+++ src/routes/GenericRoutes.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import GeneralLayout from "~/layouts/GeneralLayout";
 import Dashboard from "~/pages/Dashboard";
-import Patients from "~/pages/patients/Patients";
+import Diagnoses from "~/pages/diagnoses/Diagnoses";
 import PatientsHistory from "~/pages/PatientsHistory";
 import Settings from "~/pages/Settings";
 import { AuthProvider } from "~/pages/login/context/AuthContext";
 import UserContextProvider from "~/pages/login/context/UserContext";
-import DiagnosisProvider from "~/pages/patients/context/DiagnosisProvider";
+import DiagnosisProvider from "~/pages/diagnoses/context/DiagnosisProvider";
 
 const GenericRoutes = {
   path: "",
@@ -20,7 +20,7 @@ const GenericRoutes = {
       path: "patients",
       element: (
         <DiagnosisProvider>
-          <Patients />
+          <Diagnoses />
         </DiagnosisProvider>
       ),
     },
